requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
websockets
pandas==2.2.3
python-multipart==0.0.19
pydantic==2.10.4
google-generativeai==0.8.3
python-dotenv==1.0.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
isodate
transformers
torch
matplotlib
seaborn
scikit-learn
numpy
